- Change crypto includes to reduce compilation times - Update cxxfolders dependency - Only statically link libc if forced OR using MingW - Add Crypto++ dependency for AES and others - Added AES helper library - P2P::Node now queues incoming data packets into messages() - Added additional send() and reply() methods to P2P::Node - Adjusted rounding issue in network transaction fee calculation - Adjusted network transaction fee per byte to account for the newer transaction structure(s)

+++ src/common/network_fees.cpp
@@ -20,7 +20,8 @@ namespace Common::NetworkFees
         transaction_size -= Configuration::Transaction::Fees::BASE_CHUNK_SIZE;
 
         // calculate the number of chunks required and round up
-        const auto chunks = uint64_t(ceil(transaction_size / Configuration::Transaction::Fees::CHUNK_SIZE));
+        const auto chunks =
+            uint64_t(ceil(double(transaction_size) / double(Configuration::Transaction::Fees::CHUNK_SIZE)));
 
         // calculate the cost of the non-free chunks
         const auto chunk_cost = chunks * Configuration::Transaction::Fees::CHUNK_FEE;
@@ -82,7 +83,7 @@ namespace Common::NetworkFees
         const auto discount_percentage = calculate_transaction_discount(leading_zeros);
 
         // determine the resulting fee discount
-        const auto fee_discount = uint64_t(floor(base_transaction_fee * discount_percentage));
+        const auto fee_discount = uint64_t(floor(float(base_transaction_fee) * discount_percentage));
 
         // determine the discounted fee
         const auto discounted_fee = base_transaction_fee - fee_discount;
